+++ src/number/index.ts
@@ -1,1 +1,2 @@
+export * from './integer.js'
 export * from './number.js'